Moving accounts page to use uid
