+++ src/pages/Accounts/components/Account/index.tsx
@@ -4,10 +4,10 @@ import axios, { Response } from 'redaxios';
 import {
   Account as AccountModel,
   AccountResponse
-} from '../../../models/Account';
-import { useNotificationDispatch } from '../../../providers/NotificationProvider';
-import { ACCOUNT_FUNCTIONS } from '../../../utils/Constants/APIConstants';
-import { NOTIFICATION_THEME_FAILURE } from '../../../utils/Constants/ThemeConstants';
+} from '../../../../models/Account';
+import { useNotificationDispatch } from '../../../../providers/NotificationProvider';
+import { ACCOUNT_FUNCTIONS } from '../../../../utils/Constants/APIConstants';
+import { NOTIFICATION_THEME_FAILURE } from '../../../../utils/Constants/ThemeConstants';
 import AccountItem from './AccountItem';
 import AccountsLoading from './AccountsLoading';
 
@@ -20,25 +20,25 @@ const Account: React.FC<AccountProps> = ({ account }) => {
   const {
     data: accountResponse,
     isLoading: loadingAccount,
-    isFetching: fetchingAccount,
-    error: accountError
+    isFetching: fetchingAccount
   } = useQuery<Response<AccountResponse[]>, Response<Error>>(
     account.id,
     async () =>
       await axios.get<AccountResponse[]>(
         `${ACCOUNT_FUNCTIONS}/${account.accessToken}`
-      )
-  );
-
-  if (accountError) {
-    notificationDispatch({
-      type: 'ADD_NOTIFICATION',
-      payload: {
-        content: `Request failed due to ${accountError.statusText}`,
-        theme: NOTIFICATION_THEME_FAILURE
+      ),
+    {
+      onError: (accountError) => {
+        notificationDispatch({
+          type: 'ADD_NOTIFICATION',
+          payload: {
+            content: `Request failed due to ${accountError.statusText}`,
+            theme: NOTIFICATION_THEME_FAILURE
+          }
+        });
       }
-    });
-  }
+    }
+  );
 
   return (
     <div>
